Use int in VM, since it's enough

+++ vm/ic.c
@@ -1,8 +1,6 @@
 #include <stdlib.h>
 #include <stdio.h>
-#include <stdint.h>
 #include <stdbool.h>
-#include <inttypes.h>
 #include <string.h>
 
 #ifdef _WIN32
@@ -12,36 +10,36 @@
 
 #include "profile.h"
 
-int64_t *mem = NULL;
-int64_t mem_size = 0;
+int *mem = NULL;
+int mem_size = 0;
 
-int64_t ip = 0;
-int64_t rb = 0;
+int ip = 0;
+int rb = 0;
 
-void resize_mem(int64_t addr) {
+void resize_mem(int addr) {
     profile_addr(addr);
 
     if (addr >= mem_size) {
-        int64_t old_mem_size = mem_size;
+        int old_mem_size = mem_size;
         while (addr >= mem_size) mem_size <<= 1;
-        mem = (int64_t *)realloc(mem, mem_size * sizeof(int64_t));
-        memset(mem + old_mem_size, 0, (mem_size - old_mem_size) * sizeof(int64_t));
+        mem = (int *)realloc(mem, mem_size * sizeof(int));
+        memset(mem + old_mem_size, 0, (mem_size - old_mem_size) * sizeof(int));
     }
 }
 
-int64_t get_mem(int64_t addr) {
+int get_mem(int addr) {
     resize_mem(addr);
     return mem[addr];
 }
 
-void set_mem(int64_t addr, int64_t val) {
+void set_mem(int addr, int val) {
     resize_mem(addr);
     mem[addr] = val;
 }
 
 const int MODE_MUL[] = { 100, 1000, 10000 };
 
-int64_t get_param(int idx) {
+int get_param(int idx) {
     int mode = get_mem(ip) / MODE_MUL[idx] % 10;
     switch (mode) {
         case 0: // position mode
@@ -51,12 +49,12 @@ int64_t get_param(int idx) {
         case 2: // relative mode
             return get_mem(rb + get_mem(ip + idx + 1));
         default:
-            fprintf(stderr, "mode error: ip %" PRId64 " idx %d\n", ip, idx);
+            fprintf(stderr, "mode error: ip %d idx %d\n", ip, idx);
             exit(1);
     }
 }
 
-void set_param(int idx, int64_t val) {
+void set_param(int idx, int val) {
     int mode = get_mem(ip) / MODE_MUL[idx] % 10;
     switch (mode) {
         case 0: // position mode
@@ -66,12 +64,12 @@ void set_param(int idx, int64_t val) {
             set_mem(rb + get_mem(ip + idx + 1), val);
             break;
         default:
-            fprintf(stderr, "mode error: ip %" PRId64 " idx %d\n", ip, idx);
+            fprintf(stderr, "mode error: ip %d idx %d\n", ip, idx);
             exit(1);
     }
 }
 
-void run(int64_t (*get_input)(), void (*set_output)(int64_t)) {
+void run(int (*get_input)(), void (*set_output)(int)) {
     while (true) {
         profile_inst();
 
@@ -87,7 +85,7 @@ void run(int64_t (*get_input)(), void (*set_output)(int64_t)) {
                 ip += 4;
                 break;
             case 3: { // in
-                int64_t value = get_input();
+                int value = get_input();
                 if (value == EOF) {
                     fprintf(stderr, "no more inputs\n");
                     exit(1);
@@ -97,7 +95,7 @@ void run(int64_t (*get_input)(), void (*set_output)(int64_t)) {
                 break;
             }
             case 4: { // out
-                int64_t value = get_param(0);
+                int value = get_param(0);
                 ip += 2;
                 set_output(value);
                 break;
@@ -131,17 +129,17 @@ void run(int64_t (*get_input)(), void (*set_output)(int64_t)) {
             case 99: // hlt
                 return;
             default:
-                fprintf(stderr, "opcode error: ip %" PRId64 " oc %d\n", ip, oc);
+                fprintf(stderr, "opcode error: ip %d oc %d\n", ip, oc);
                 exit(1);
         }
     }
 }
 
-int64_t get_input() {
+int get_input() {
     return getc(stdin);
 }
 
-void set_output(int64_t val) {
+void set_output(int val) {
     putc(val, stdout);
     fflush(stdout);
 }
@@ -154,17 +152,17 @@ int main(int argc, char **argv) {
     profile_init();
 
     mem_size = 64;
-    mem = (int64_t*)malloc(mem_size * sizeof(int64_t));
-    memset(mem, 0, mem_size * sizeof(int64_t));
+    mem = (int*)malloc(mem_size * sizeof(int));
+    memset(mem, 0, mem_size * sizeof(int));
 
     FILE *input = fopen(argv[1], "rt");
 
     int idx = 0;
     char ch = ',';
-    int64_t num = 0;
+    int num = 0;
 
     while (ch == ',') {
-        if (fscanf(input, "%" SCNd64 "%c", &num, &ch) != 2) {
+        if (fscanf(input, "%d%c", &num, &ch) != 2) {
             fprintf(stderr, "parse error");
             exit(1);
         }
